bare_trace: abstract some of the basic FTrace functionality to a generic BareTrace class
Part of the functionality in FTrace is generic.  In order to make the
rest of trappy less dependent on FTrace, move the generic parts to a new
BareTrace class and inherit from it.
