+++ trappy/ftrace.py
@@ -22,6 +22,7 @@ import os
 import re
 import pandas as pd
 
+from trappy.bare_trace import BareTrace
 from trappy.utils import listify
 
 def _plot_freq_hists(allfreqs, what, axis, title):
@@ -42,7 +43,7 @@ def _plot_freq_hists(allfreqs, what, axis, title):
         trappy.plot_utils.plot_hist(allfreqs[actor], ax, this_title, "KHz", 20,
                              "Frequency", xlim, "default")
 
-class FTrace(object):
+class FTrace(BareTrace):
     """A wrapper class that initializes all the classes of a given run
 
     - The FTrace class can receive the following optional parameters.
@@ -108,11 +109,10 @@ class FTrace(object):
 
     def __init__(self, path=".", name="", normalize_time=True, scope="all",
                  events=[], window=(0, None), abs_window=(0, None)):
-        self.name = name
+        super(FTrace, self).__init__(name)
+
         self.trace_path, self.trace_path_raw = self.__process_path(path)
-        self.class_definitions = self.dynamic_classes.copy()
-        self.basetime = 0
-        self.normalized_time = False
+        self.class_definitions.update(self.dynamic_classes.items())
 
         self.__add_events(listify(events))
 
@@ -124,7 +124,6 @@ class FTrace(object):
             self.class_definitions.update(self.thermal_classes.items() +
                                           self.sched_classes.items())
 
-        self.trace_classes = []
         for attr, class_def in self.class_definitions.iteritems():
             trace_class = class_def()
             setattr(self, attr, trace_class)
@@ -132,7 +131,7 @@ class FTrace(object):
 
         self.__parse_trace_file(window, abs_window)
         self.__parse_trace_file(window, abs_window, raw=True)
-        self.__finalize_objects()
+        self.finalize_objects()
 
         if normalize_time:
             self.normalize_time()
@@ -205,25 +204,6 @@ class FTrace(object):
         with open(raw_trace_output, "w") as fout:
             fout.write(raw_out)
 
-    def get_duration(self):
-        """Returns the largest time value of all classes,
-        returns 0 if the data frames of all classes are empty"""
-        durations = []
-
-        for trace_class in self.trace_classes:
-            try:
-                durations.append(trace_class.data_frame.index[-1])
-            except IndexError:
-                pass
-
-        if len(durations) == 0:
-            return 0
-
-        if self.normalized_time:
-            return max(durations)
-        else:
-            return max(durations) - self.basetime
-
     @classmethod
     def register_class(cls, cobject, scope="all"):
         """Register the class as an Event. This function
@@ -244,85 +224,6 @@ class FTrace(object):
             cls.dynamic_classes[cobject.name] = cobject
         else:
             getattr(cls, scope + "_classes")[cobject.name] = cobject
-
-    def get_filters(self, key=""):
-        """Returns an array with the available filters.
-
-        :param key: If specified, returns a subset of the available filters
-            that contain 'key' in their name (e.g., :code:`key="sched"` returns
-            only the :code:`"sched"` related filters)."""
-        filters = []
-
-        for cls in self.class_definitions:
-            if re.search(key, cls):
-                filters.append(cls)
-
-        return filters
-
-    def normalize_time(self, basetime=None):
-        """Normalize the time of all the trace classes
-
-        :param basetime: The offset which needs to be subtracted from
-            the time index
-        :type basetime: float
-        """
-
-        if basetime is not None:
-            self.basetime = basetime
-
-        for trace_class in self.trace_classes:
-            trace_class.normalize_time(self.basetime)
-
-        self.normalized_time = True
-
-    def add_parsed_event(self, name, dfr, pivot=None):
-        """Add a dataframe to the events in this trace
-
-        This function lets you add other events that have been parsed
-        by other tools to the collection of events in this instance.  For
-        example, assuming you have some events in a csv, you could add
-        them to a trace instance like this:
-
-        >>> trace = trappy.FTrace()
-        >>> counters_dfr = pd.DataFrame.from_csv("counters.csv")
-        >>> trace.add_parsed_event("pmu_counters", counters_dfr)
-
-        Now you can access :code:`trace.pmu_counters` as you would with any
-        other trace event and other trappy classes can interact with
-        them.
-
-        :param name: The attribute name in this FTrace instance.  As in the example above, if :code:`name` is "pmu_counters", the parsed event will be accessible using :code:`trace.pmu_counters`.
-        :type name: str
-
-        :param dfr: :mod:`pandas.DataFrame` containing the events.  Its index should be time in seconds.  Its columns are the events.
-        :type dfr: :mod:`pandas.DataFrame`
-
-        :param pivot: The data column about which the data can be grouped
-        :type pivot: str
-
-        """
-        from trappy.base import Base
-        from trappy.dynamic import DynamicTypeFactory, default_init
-
-        if hasattr(self, name):
-            raise ValueError("event {} already present".format(name))
-
-        kwords = {
-            "__init__": default_init,
-            "unique_word": name + ":",
-            "name": name,
-        }
-
-        trace_class = DynamicTypeFactory(name, (Base,), kwords)
-        self.class_definitions[name] = trace_class
-
-        event = trace_class()
-        self.trace_classes.append(event)
-        event.data_frame = dfr
-        if pivot:
-            event.pivot = pivot
-
-        setattr(self, name, event)
 
     def __add_events(self, events):
         """Add events to the class_definitions
@@ -466,11 +367,6 @@ class FTrace(object):
             fin.seek(0)
 
             self.__populate_data(fin, cls_for_unique_word, window, abs_window)
-
-    def __finalize_objects(self):
-        for trace_class in self.trace_classes:
-            trace_class.create_dataframe()
-            trace_class.finalize_object()
 
     # TODO: Move thermal specific functionality
